一、修改首页界面 二、定位不打开不给使用 三、增加camera权限申请

+++ app/src/main/java/ai/woyao/anything/bike/net/bean/params/ParamsQr.java
@@ -6,6 +6,6 @@ package ai.woyao.anything.bike.net.bean.params;
 
 public class ParamsQr {
     public String bike_id;
-    public long lat;
-    public long lng;
+    public double lat;
+    public double lng;
 }
